feat: set Methods Parameters can be nullable
The Set Methods from Client and Company classes now have its parameters accepting null. This comes
from an bug when, for example, an client don't have any city, and the Spreadsheet Reader understand
it as Null, then, in set methods, the parameter will also be null. This was corrected by accepting
null values in the set parameters.